pyproject.toml
[build-system]
requires = ["hatchling>=1.22.0"]
build-backend = "hatchling.build"

[project]
name = "llmflow-core"
version = "0.0.1"
description = "Deterministic LLM workflow engine"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Ibrahim" }]
keywords = ["llm", "workflow", "deterministic", "pipelines"]

# Runtime dependencies
dependencies = [
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "jsonschema>=4.0",
  "jinja2>=3.1",
  "typer>=0.9",
  "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
